Remove special cases for non-tuple array element access
Summary:
It isn't necessary to do this juggling with `NumT` for arrays. If the key type
isn't a number, we will error when handling the `ElemT` constraint.

This special case had the effect of causing an error if the key type is open and
never gets a lower bound, but that behavior is consistent with the rest of Flow
overall. Furthermore, creating a `NumT` like this will thwart reachability
analysis, because unreachable stuff that participates in an elem action looks
reachable (gets lower bounds) when it isn't (doesn't).

Reviewed By: gabelevi

Differential Revision: D4175122

fbshipit-source-id: e15240c209ee494b3b121b41858f09ec9f616a32