Categorize unsupported destructuring errors
Summary:
Before this diff, we used the error message "unsupported destructuring" to
describe a few different errors. It turns out that only two of those cases were
actual unsupported syntax: non-string literal object properties and expression
patterns.

There are two other cases:

1. Invalid LHS in assignment. This isn't a destructuring error, so I created a
new more explicit error type for this.

2. Non-identifier rest element in function param list. This is an internal error
(i.e., it is a parse error, but we need to handle it during analysis because the
AST is overly general.) I created a new internal error type for this case.

Reviewed By: mroch

Differential Revision: D4179138

fbshipit-source-id: 48553431743fc5bddea58268a56f07de42cb7c2e