[hack] v0.2.2 - Support for Switch/Case/Default Statement and other minor ast nodes
Summary:
Handling switch/case is fun. For an example like
  switch ($var) {
    case 1:
      $a = 1;
      return $a;
  }
The current hack grammar parses the body into a list of statements. `$a = 1;` is parsed as a child of the `case 1:` statement, and `return $a;` is a sibling to the `case 1:` statement, so I handle this by unnesting the statements and handling them as a list.

Differential Revision: D4115950

fbshipit-source-id: 24e00996b2ed58d1d23bd41c7a19b3df0bfc4837