add --all and --respect-pragma flags to coverage
Summary:
adds the same `--respect-pragma` flag that `flow check-contents` has to
`flow coverage`. when not set, `flow coverage` checks the file as if it has
`flow`, even if it doesn't. when set, it says every expression is uncovered.

**this flag is immediately deprecated.** you can use it to get the new
behavior a release or two sooner, but will need to remove it again soon.

also adds an `--all` flag to both `check-contents` and `coverage`, which is
currently a no-op. in a future version of flow, we will flip the default. so if you
want to maintain the current behavior, add `--all` to your commands now.

Reviewed By: gabelevi

Differential Revision: D4181810

fbshipit-source-id: 8f88f4053919f0acbac7cc48f8a1139073a014fb