[easy] Clean up some module/import/export APIs + fix a bug
Summary:
I'm working on `declare module` imports, and this is the first minor preface to that work.

This cleans up the naming of some APIs to be a little more clear and it fixes a bug where we were returning the `ModuleT` for the current `declare module` no matter what named module you ask for (if and when you're inferring the interior of a `declare module` body).

This bug is coincidentally unobservable at the moment because we never ask for the `ModuleT` of anything inside a `declare module "X"` body except for `"X"` itself. This will change with `declare module` imports, though

Reviewed By: gabelevi

Differential Revision: D4180828

fbshipit-source-id: 914fc41773c86cab6da5d594f9fffa26586c0e68