Discharge EvalT constraints eagerly
Summary:
In the included test, we create an EvalT for the destructuring and flow that
into the OpenT representing the variable `x`. Since `x` never gets an upper
bound, the EvalT is never discharged and we never realize a type error.

This diff takes a controversial approach to the problem: evaluate these types
eagerly, before they collect in the bounds of tvars.

Why is this OK to do? Well, consider some hypothetical `EvalT -> T` flow.
Before this change, we would wait until some `T -> U` constraint, then discharge
`EvalT -> U`. The `EvalT` becomes some type `X` and then `X -> U`.

After this change, we `EvalT -> T` becomes `X -> T`, then when `T -> U` we flow
`X -> U`, so when uppers do run into lowers, we ultimately create the same
constraints, just in a different order.

The only difference, then, is the difference that we want: EvalT types are
processed even if they never run into an upper bound.

Reviewed By: avikchaudhuri

Differential Revision: D4180176

fbshipit-source-id: 40c94ab6b0c6510c39f76e15b67ce42bb6cf5f99