Preserve union type destructured upper bounds
Summary:
As shown in the added tests, union types that pass through type destructors get
split apart, which breaks the speculative match behavior union upper bounds
usually exhibit.

This diff fixes that behavior by preserving the union through a type destructor.
Instead of evaluating the destructor for the entire union (which causes us to
hit the `(UnionT _, _)` flow and split the union up), we create new eval IDs and
map the type destructor over the branches of the union.

Reviewed By: mroch

Differential Revision: D4202202

fbshipit-source-id: 1212ca11e525a219d0742fbb594007075c698ef8