[hack] Load the deptable in the master process not the deamon
Summary:
Deptable needs to be loaded in the master process since the master process is the only process accessing the deptable. With the SQL approach, we use static variables and mem location of these variables do not get shared across processes.
This addresses SEV 142194

Also realizing that this approach requires workers to be able to access the sql table hence, I'm enabling that as well.

Reviewed By: jopamer

Differential Revision: D4194412

fbshipit-source-id: 2fc8008bab76372a95485c91477b0859b7623b23