flow EmptyT out of impossible MaybeTs
Summary:
`$NonMaybeType<T>` creates a tvar that receives the unwrapped `T'`. if
`T` is null, then we never flow anything into that tvar, which leaves the tvar
unresolved and it accepts whatever.

This diff flows `EmptyT` out of a `MaybeT` on `null` or `void`. It also
special-cases `MaybeT ~> MaybeT` to avoid breaking apart the implicit union.
Otherwise, `MaybeT (NumT) ~> MaybeT (TVAR)` results in `TVAR` having `EmptyT;
EmptyT; NumT` as lower bounds, which isn't what we want.

Depends on D4202202

Reviewed By: samwgoldman

Differential Revision: D4205909

fbshipit-source-id: 0059cd4903ce8da6f8b40d047fc5ae91bc899c61