Support `import type/typeof` inside `declare module`
Summary:
This diff adds parser + inference support for importing types from other `declare module`s.

The one caveat to this approach is that it **does not** yet support libdef `declare module`s that import from non-libdef files. To do this will require more extensive factoring of how we merge libdefs -- but supporting that is a medium-priority goal on my roadmap.

Reviewed By: avikchaudhuri

Differential Revision: D4205225

fbshipit-source-id: fa18878d7061e36f0fd00f6e5ec40125487bbbe9