refactor root stripping (part 4)
Summary:
stripping the root dir from paths in error output was a mess. in
various places, we either stripped the root when printing a path, or created new
`Loc`s with relative, already-stripped sources. Sometimes we even did both
(stripped an already-stripped loc), or had to recover an absolute path by
passing a `~root` param and prepending it to an already-stripped loc.

This diff pushes all root stripping to the places where we print locations. A
`Loc` should always have an absolute, unstripped source.

Aside from preventing errors like one I fixed recently where we didn't print
context lines for lib files in errors because we couldn't find a relative path
as an absolute path, this should also make it safer to compare two Locs.

Reviewed By: samwgoldman

Differential Revision: D4217102

fbshipit-source-id: 6defacbdc7814542ab9a7bf3626f759480aafeb1