`flow ls` improvements
Summary:
In an ideal world, you would be able to ask flow "what are all the
roots that care about file foo.js?". This is useful for integrations that need
to run all relevant flow checks whenever foo.js changes. However, this is
difficult, since it requires knowledge of every .flowconfig on the machine.

An easier question to answer is "Does this root care about foo.js?" This diff introduces
two new flags:

* `flow ls --all`, which includes lib files and explicitly ignored files too
* `flow ls --explain`, which gives what kind of file each file is

The six reasons are

  * Implicitly a lib (in the flow-typed directory)
  * Explicitly a lib (in the libs section of the .flowconfig)
  * Implicitly ignored (not in the root dir and not explicitly included)
  * Explicitly ignored (in the ignores section of the .flowconfig)
  * Implicitly included (not ignored and in the root dir)
  * Explicitly include (not ignored and in the includes section of the .flowconfig)

Reviewed By: mroch

Differential Revision: D4160418

fbshipit-source-id: f89290ce80d6899176873242dd51e5aa913bd301