FormData definitions
Summary:
Just filled out the definitions for [`FormData`](https://developer.mozilla.org/en/docs/Web/API/FormData).

I've updated the test which was failing from line numbers changing, and I've added a test for `FormData` (there was no `bom` dir - I assume this is the correct place for them?)

One concern there might be is I've tightened the types (on the sole existing method) from `any` to `string`, which seems to be correct as a string _is_ the correct value to give to these functions. Give it an object and you get `"[Object object]"` back.

``` diff
- append(name: any, value: any, blobName?: string): void;
+ append(name: string, value: string, blobName?: string): void;
```
Closes https://github.com/facebook/flow/pull/2471

Reviewed By: samwgoldman

Differential Revision: D3868288

Pulled By: mroch

fbshipit-source-id: b0a15fbed8f74c08357ec821f9e6c8fec7c58e41