Fix build issue with LZ4
Summary:
Because the Makefile builds C files in the order given by `NATIVE_C_FILES`, and the ordering of `$(wildcard ...)` is non-deterministic, it's possible for the build to compile C files from LZ4 before files they `#include` are present.

This was happening to our build, causing a failure where building `lz4hc.c` would be attempted before `lz4.c` was even present. The solution proposed is a hack. A better solution would probably be an explicit target for LZ4 that uses it's Makefile, but for the moment I just wanted to highlight the issue and show _a_ solution for others who might be experiencing this problem.
Closes https://github.com/facebook/flow/pull/2879

Reviewed By: mroch

Differential Revision: D4226797

Pulled By: gabelevi

fbshipit-source-id: 73983440a11f272e19f9b7083f8bb05ff97b49ea