Add parsing time to js_test_files.js
Summary:
This is a script that helps test the JS version of the Flow parser.
It's useful to see how long the Flow parser takes to run. I can then run

  cd ~/flow/src/parser/
  find -H ~/my-repo -name "*.js" | head -n 1000 | xargs ./tools/js_test_files.js -v

Which parses the first 1000 js files in my repo and outputs how long parsing
took.

The parse time excludes IO time.

Reviewed By: samwgoldman

Differential Revision: D4245458

fbshipit-source-id: bc2c7a914acaf8612c80ca4f1c8404feb75869ed