[website] handle errors when running flow better
Summary:
using `popen3` was wrong. `stdout` unblocked when stuff was written to
stderr, so `stdout.gets` returned empty string, which caused a JSON parse error
that manifested as a "nil can't be coerced to String" error.

now, `capture3` returns both stderr and stdout (and `--quiet` means we shouldn't
print to stderr), and we handle any errors more gracefully.

Reviewed By: gabelevi

Differential Revision: D4277939

fbshipit-source-id: 057b3f540c3183511277f807db2a482340180046