Make room for object type spread properties in the AST
Summary:
Object types don't support a spread property type, but we want them to. This
diff just adds a new AST node. Parsing and type checking support will be added
in subsequent diffs.

Note that the estree output is backwards compatible, as long as object type
spread properties are not used. Anything that depends on our AST should be
upgraded to support the new `ObjectTypeSpreadProperty` node.

Reviewed By: mroch

Differential Revision: D4268699

fbshipit-source-id: 542ef9f90d4eed7474ea46acad6d7f2b33b88d0d