fix exception when using - as a filename
Summary:
we use the path `-` when reading from stdin in `check-contents` when a
filename isn't set. the check for whether a path is relative only really matters
if we have to read the file from the filesystem, so moved it to allow any
filename if it matches the fake filename of the stdin stream (whether `-` or
some other potentially made-up path).

Reviewed By: jeffmo

Differential Revision: D4277902

fbshipit-source-id: b716da1240b06cf08805abedbc3bdc49a17df197