don't allow directional errors to swap uppers and lowers
Summary:
these errors are all directional, meaning that they don't make sense if
the upper and lower reasons are switched. for example, it would be bad to swap
the reasons on FunReturn which is phrased, "This type is incompatible with the
expected return type of".

strictly speaking, it's ok to swap the locations in `Addition`, but given that
we always reposition both sides of an addition, we should never end up in a
scenario where we'd need `ordered_reasons` to flip anything (one in a lib, for
example). so including it for consistency.

Reviewed By: samwgoldman

Differential Revision: D4287638

fbshipit-source-id: 37f4854d9baa1961e993e26140f2a12509fa8fb8