include "extra" info in Error.compare
Summary:
errors are de-duped using a `Set` which uses `compare` to decide
whether something's in the set already.

not comparing the `extra` field on errors means that two errors could
incorrectly dedupe if they have the same reasons but, for example, different
union mismatches.

I ran into this while working on nested object error messages. the top-level
error is that two objects are incompatible, but they differ in the extra fields
where we describe why they're incompatible.

Reviewed By: samwgoldman

Differential Revision: D4286631

fbshipit-source-id: 28b137718921f903709bedcd06ca5d8ca336ced9