recheck opts
Summary:
Currently, any edit to a file causes rechecking of its recursively dependent
files. This is usually overkill: a file must be rechecked only if the signatures
of any of the files it directly depends on have changed.

This diff tracks when the signature of a file hasn't changed, and modifies the
stream of work generated during merge to be aware of this fact.

Reviewed By: samwgoldman

Differential Revision: D4219850

fbshipit-source-id: 7779fe93b90d0d78324a8bcf79b81910afcba9ae