find refs: locals
Summary:
Find refs for locals.

Usage:
`flow find-refs file line column`

If `file:line:column` is pointing at `x` in a read expression `x`, a write expression `x = ...`, or a definition site `var/let/const x ...` of a local variable `x`, then this command should return the locations of all reads, writes, and the definition site of `x`.

This feature can be used in an IDE, e.g., to highlight all occurrences of a local variable when the cursor is on one of them, and eventually to implement renaming (although how to rename shorthand object properties remains TBD).

In the future, this feature will be generalized to work for object / class-instance properties as well, but that would require some work on alias analysis first.

Reviewed By: samwgoldman

Differential Revision: D4127493

fbshipit-source-id: 02ce77bf862c196dcef76beede655579829e3ff8