skip checking an object against itself
Summary:
when an ObjT flows into an ObjT with the same property id, then it's
the same object and we know that all of its props are compatible. Also, since
the propmap ID is allocated when the ObjT is created, and the only thing we
should be changing on an ObjT is its reason, it should be safe to just compare
IDs rather than all the other fields on the object.

Reviewed By: avikchaudhuri

Differential Revision: D4321066

fbshipit-source-id: 563b30141ed5478f8b80ac7537bafa3548c88212