[hack] Create Ide_message module
Summary:
Some refactoring in preparation for new API support. This, along with few follow up diffs, is going to replace a bunch of bad sphagetti code from IdeJson and IdeJsonUtils utils with something more structured and extensible.

Treat it as ground up rewrite - in the final diff I'll just remove IdeJson / IdeJsonUtils and replace all of their usages with this.

Reviewed By: jamesjwu

Differential Revision: D4297776

fbshipit-source-id: ea0b43a0cda24bc9fb564a02de98c62f0455a85c