improve error when adding mixed
Summary:
it's irrelevant what the other side of an addition is when one side is
`mixed`. saying that `mixed` can't be added to `something else` is confusing
because it implies that you could change `something else` to make it work.

instead, use a one-sided error that tries to explain why `mixed` in addition is
bad, in a relatively succinct way.

Reviewed By: gabelevi

Differential Revision: D4309592

fbshipit-source-id: 14ecbb05d84b0f9b43580b5ccf19eafbf4a9da9a