recheck leaders
Summary:
We were marking directly dependent files as needing rechecks, but if the
directly dependent files have a cycle and a different leader is picked for that
cycle than the marked file, the cycle was skipped during recheck.

This is obviously wrong, so now we mark leaders for rechecking whenever any
files in their components are marked.

A better solution eventually would be to have a different type for leaders so
that this kind of "type confusion" is caught by the compiler.

Reviewed By: gabelevi

Differential Revision: D4325324

fbshipit-source-id: f8e916e5481c7569a903222d90dd0d51dc5df38f