fix declaration_files_incremental_haste test
Summary:
this test was attempting to `mv` files that didn't exist. as long as
the test passed, the errors from `mv` were hidden because we hide `stderr`. if
the expected errors change, the test failure is confusing because then it shows
`stderr`.

this fixes the bad moves, and also redirects the `mv` output to stdout so it
shows up in the error output instead of being swallowed the future.

Reviewed By: samwgoldman

Differential Revision: D4334015

fbshipit-source-id: bc465b2444e61efed461c2e9dc24f18c8b6d3a2d