add ability to skip esprima3 tests
Summary:
in order to enable the esprima3 test suite, this adds the ability to
skip tests that we don't want to run yet.

to avoid touching any of the original files (to avoid merge conflicts), this
lets you add a file ending in `.skip`, where the contents is the reason for
skipping the test. it'll print like this:

```
=== ES6/identifier ===
[-] SKIP: dakuten_handakuten - TODO: parse errors
[-] SKIP: escaped_all - TODO: parse errors
[-] SKIP: escaped_math_alef - TODO: parse errors
[-] SKIP: escaped_math_dal_part - TODO: parse errors
[-] SKIP: escaped_math_kaf_lam - TODO: parse errors
```

Reviewed By: samwgoldman

Differential Revision: D4332450

fbshipit-source-id: f297a436a3f70c3b3c07fd5904a46a0801552f37