make sentinel prop tests consistent with other conditionals
Summary:
we intentionally allow lookups of props that don't exist when inside a conditional, e.g. `if (foo.bar)` or `if (foo.bar === undefined)`. however, we currently forbid them in sentinel property tests of the form `if (foo.bar === 'string')` (same for other literals). this appears to be an oversight.

this diff loosens sentinel prop checks to allow non-existant props, like other conditionals.

Reviewed By: samwgoldman

Differential Revision: D4328249

fbshipit-source-id: 57ce5e2f66e8ed8b92327b7c5f24b2e89769eb83