refine sentinel props when types don't match
Summary:
when we do a sentinel prop test, and we know that the type of the RHS
doesn't match the type of the prop, we can prune it.

currently, we only prune when the RHS is a literal (a known value like 'foo',
rather than any string) and the prop is a different literal of the same type.

now, we also prune when the RHS is a literal, and the prop is a different type.

NOTE: we don't prune when the RHS is a non-literal even when the type is wrong,
since we don't send all prop equality tests through the sentinel prop flow (see the
last test in tagged_union.js).

Reviewed By: samwgoldman

Differential Revision: D4307843

fbshipit-source-id: a1ba9cc54f78f2d8012d2cccd9824e9d0c060906