hash tvar and propmap ids as well
Summary:
It turns out that only hashing nominal (class) ids is not sufficient. While
computing signatures we avoid visiting the same tvars and propmaps again to
avoid looping, but this can be exploited to defeat the hashing scheme: different
tvars and propmaps that have already been visited look the same. To fix this, we
hash their ids when we visit them again.

Reviewed By: mroch

Differential Revision: D4332270

fbshipit-source-id: fa65d0876f1064fbce7c267b3a7fa14fdb1a867d