don't relativize class ids that are imported when hashing
Summary:
Previously, a class id could change because a file was rechecked, yet not force
a change in the hash of a downstream file that imports it and re-exports
it. This might cause type confusion for files further downstream because there
are now two copies of the same class id, one new and one old.

This diff fixes. (Recall that some relativization is still desirable, otherwise
we'd force downstream files to trivially recheck every time a file defining a
class is changed.)

Reviewed By: mroch

Differential Revision: D4337420

fbshipit-source-id: 20f39b1b32e5e3087c53904bc2007832382b87d5