sort components to pick the same leader, justifying skip
Summary:
Previously skipping a file on recheck was dangerous: it could cause a
crash. Consider a cycle with two files, f1 and f2. We pick f1 the leader during
init, then make an unrelated change and somehow pick f2 the leader during
recheck. (This was possible because the dependency graph could be walked in any
order.) Then skipping f2 would amount to trying to fall back on an existing
signature, but that signature would be associated with f1.

This diff sorts components so that the leader of the same component is always
the same. This makes skipping safe.

Reviewed By: samwgoldman

Differential Revision: D4339309

fbshipit-source-id: eceef806f2f5353075394f01463491a23293810e