replace component sorting with leader update
Summary:
D4339309 tried to fix merge crashes during rechecks by sorting components,
thereby ensuring components have the same leaders no matter how they are
discovered. Unfortunately, this strategy currently turns out to be problematic:
sorting perturbs the order in which files of a component are checked, which
somehow (probably interacting with other bugs) causes spurious error diffs on
rechecks.

This diff implements an alternative to sorting, where the system adapts when
the leader of a component changes, either installing the new leader or reusing
the old leader. It also makes the handling of leaders similar to the handling of
sig contexts and sig hashes, thereby simplifying some invariants.

Reviewed By: jeffmo

Differential Revision: D4345135

fbshipit-source-id: bcb1c02c0449e6fcb65134e6b1796c7661367801