Treat single-element array literals as still open
Summary:
Previously, we considered the type of `[e]` where `e:T` to be exactly
`Array<T>`. However, the type of `[e0,e1]` where `e0:T0` and `e1:T1` would be
`Array<T>` where `T` is open and `T0 < T` and `T1 < T`.

This diff makes the single-element case mirror the multiple-element case. This
makes sense from a consistency standpoint -- why should `[1]` and `[1, 2]`
behave differently?

The openness of the element type has an effect on the new union type stuff, so
there is some related error diff, but I feel this behavior is better on the
balance.

Reviewed By: gabelevi

Differential Revision: D4353648

fbshipit-source-id: fb213debb7146caebbe6f8892dc1916f696e97b4