hash property kind
Summary:
During recheck, we create a pared down signature of the exports of a module. If
the signature does not change when the file is touched or a dependency changes,
we don't need to recheck dependents.

Before this diff, the signature did not include some property information, in
particular variance. This means that a diff that adds variance, or changes a
field to a getter, would not cause dependents to be rechecked.

Reviewed By: mroch

Differential Revision: D4367782

fbshipit-source-id: 6265d9b6e00794da551a13118b2d97f5f5759ae1