nested property compat errors
Summary:
when we flow an `ObjT` into another `ObjT`, we flow each prop into the
corresponding prop of the other object. sometimes those props are incompatible,
and error. the locations of the types that error are the def sites of the object
types themselves, not their use sites or the place where the objects collided.
this produces hard-to-debug "errors at a distance"; sometimes both sides of the
error can be in a different file from the one responsible.

instead, we can error when the objects themselves are incompatible, and explain
that it's because a certain property didn't match. this positions the error at
the place where the objects collide, which is much more likely to be actionable.

this is achieved by storing the reasons of the objects in the `use_op` that we
pass when we flow the props together, as in

```
(* pseudo-ish code *)
flow cx trace (l_prop, UseT (PropertyCompatibility (l_obj, u_obj), u_prop)
```

when `l_prop` errors against `u_prop`, we can instead say that `l_obj` and
`u_obj` are incompatible, because of `l_prop` and `u_prop`.

Reviewed By: gabelevi

Differential Revision: D4322000

fbshipit-source-id: d4babb5e0a49786da0e13ef56c5efae9a8361a15