Save server logs for tests with a .testconfig
Summary:
The runtests script uses flow start, which outputs to a log file. Any printf
debugging gets directed to the log file, but we can't see where it is. This diff
uses the --log-file flag, so the log file is saved to the test directory.

The log file is created and cleaned up by the same rules as the .exp, .out,
.diff files.

Reviewed By: gabelevi

Differential Revision: D4372892

fbshipit-source-id: 01ea7b3c25c6fb6a7258f5a1eb9918ccbe25db52