nested prop compat errors (part 2)
Summary:
a followup to D4322000

when an object type (vs an object literal) flows into another object type, we
hit a different codepath that needs to be aware of `use_op` to get nested
errors.

the props go through unification, so `__unify` needs to thread the use_op
around. this diff catches the biggest case; I need to come back when I'm not on
PTO to hit the other cases like obj -> obj unification.

Reviewed By: gabelevi

Differential Revision: D4374943

fbshipit-source-id: 0c92e4e97ec3c44f6fcb8b03f7e611a9cc932083