[easy] minor parser refactoring
Summary:
I was looking into removing `Loc.none`, but it's being used for
normalized reasons so I don't think we can kill it.

But I came across these callers, which I don't think are legit. If the lists are
somehow empty, an invariant is being violated and we should throw instead of
returning a bogus location that might end before it starts.

Reviewed By: gabelevi

Differential Revision: D4379029

fbshipit-source-id: 0583b41fb030d0026db8b3d2e6a3ed3ae668a6be