error on computed getters and setters
Summary:
computed getters and setters are currently treated like normal computed
properties, as props whose values are functions, because we handle computed
props before getter/setter-ness. oops!

this diff makes it so that a computed get/set errors, since we don't have the
mechanisms to support them properly yet. they error even when
`enable_unsafe_getters_and_setters` is enabled, because the problem is that
they're computed, not that they're unsafe.

Reviewed By: samwgoldman

Differential Revision: D4388624

fbshipit-source-id: 01c71de1f7047f91cd2eb4c06bfe7777cc0cfc1c