Switch ./tool to use node 6 in buck
Summary:
When ./tool was outputing lots of stdout (like when the dom test
fails), it would occasionally truncate the stdout, print the stderr that comes
after the stdout, and exit with code 0 (it should have been code 1). I have no
idea why it does this. Nothing I did seemed to fix it, other than using node 6
instead of node 4. *shrug*

Reviewed By: mroch

Differential Revision: D4369296

fbshipit-source-id: d0358b8fcbb5f30752e2066511b3ad79d67aa17e