Split funtype into funtype and funcalltype
Summary:
As a shortcut, `FunT` was sharing its `funtype` record with the
"function call" types (`CallT`, `ApplyT`, `BindT`, etc). This is undesirable
for a few reasons

* A function parameter list (the params a function takes) is different than a
  function argument list (the arguments passed to a function). The former can
  have a single rest element at the end, while the latter can have multiple
  spread elements.
* There are a bunch of fields that `FunT` needs which the others don't

So I split them up. This also let me separate out the rest_param from the list
of params, which is structurally cleaner than including it in the list of
params.

This was mainly a OCaml compiler directed change. However,
I also grep'd for `param_tlist` to figure out which code also would care about
the `rest_param` field.

Reviewed By: samwgoldman

Differential Revision: D4372185

fbshipit-source-id: 2b7b3bde090e650b81a52cf67efa0bbdbeb16a5c