[hack] Rename vector and map types to vector_array and map_array
Summary:
The nodes for array<t> and array<k, v> were VectorType and MapType, respectively, following the specification. But we need to implement the vec<t> type, which is sensibly named VectorType.  So I've renamed them VectorArrayType and MapArrayType respectively.

That's this diff; in subsequent diffs I'll add nodes for the new intrinsic types, and then parse them.

Reviewed By: KendallHopkins

Differential Revision: D4381356

fbshipit-source-id: 2840804a6b0ca6ec30158b917f5007ec7da4a834