[hack] Include trivia that leads the end-of-file in full fidelity parse tree
Summary:
In the FFP every token has "leading" and "trailing" trivia; there has to be a line which divides them and we arbitrarily say that it is at the first newline.  So

      $x = 123;    /* spaces and this comment and the newline are trailing the semi */
      /* spaces and this comment are leading the return */ return 456;
    } // spaces and this comment and newline are trailing the }
    // This comment is leading the end-of-file.

We were losing from the tree any trivia that was past the trailing trivia of the last "real" token. So we make a magical "end of file" token and an "end of file" parse tree node, and stick it at the end. However we *only* do this in cases where there is trivia leading the end of file.

Reviewed By: KendallHopkins

Differential Revision: D4384631

fbshipit-source-id: 3b5a13cbafebc71539c1724d3560a2c2d77bbaf5