refactor parallel lists of failed parses into tuple
Summary:
iirc, we were passing around parallel lists of filenames that failed to
parse and the parse errors for each file because it was cheaper than returning a
set/map of the errors. I think we could've returned a single list of tuples for
about the same cost, and in the meantime we've changed one of the lists to be a
tuple anyway. this combines the two lists into a single tuple.

Reviewed By: gabelevi

Differential Revision: D4396062

fbshipit-source-id: fc32f2753a059addbf949fe0e3e2c51b52188fc1