implements
Summary:
A class declaration can specify interfaces which it aims to implement. Flow will
check that the implementation satisfies all declared interfaces at declaration
time.

Since we already check structural compatibility at the declaration site, it's
not necessary to perform the same structural check later on. If an instance
flows into an interface upper bound which is in its implements declaration, we
short-circuit the check based on the nominal class id for that interface.

Only interface types are allowed in the implements declaration currently. It may
be desirable to allow other types in the implements clause, but doing so is left
for future work.

Only class declarations are allowed to specify an implements clause. Supporting
implements for `declare class` is left for future work (it doesn't even parse).

Fixes https://github.com/facebook/flow/issues/2376

Reviewed By: gabelevi

Differential Revision: D4386573

fbshipit-source-id: cca153d5afab8ff23ecd60c9fe940218c7e4d398