hash file names in cycles
Summary:
Suppose that some files in a cycle are renamed (but the leader is not, and
nothing else changes). Earlier the sig hash of the leader wouldn't change, which
would cause entries for the new files to remain "unchanged" as well:
technically, empty.

Including the set of files forming a cycle in the sig hash fixes the
problem. (The modules exported by those files are already part of the sig hash,
and were included to solve a related problem.)

Reviewed By: gabelevi

Differential Revision: D4396462

fbshipit-source-id: 68d1461296186cf37146617908297533abcd3d81