$ReadOnlyArray<+T> - a supertype for all tuples and arrays
Summary:
After the recent strict tuple change, there's no supertype for all
tuples. This diff changes that by doing the following

* Renames `Array$Tuple` to `$ReadOnlyTuple`. This is more concise and looks better.
* Makes `Array` extend `$ReadOnlyTuple`
* Creates a third `ArrT` variant called `ROTupleAT elemt`, which is the
  supertype for all tuples and arrays which have elements that are subtypes of elemt.
* The type annotation `$ReadOnlyTuple<number>` creates a `ReadOnlyTuple NumT`

`$Tuple<string, number>` was basically a broken version of `[string, number]`
from back before we added the syntax for tuples. It's not documented or really
used. So I'm removing it.

Reviewed By: samwgoldman

Differential Revision: D4365960

fbshipit-source-id: b683fe50107ce74096349a3d712a34b2ab2a63a6