hash reasons
Summary:
We were not hashing locations in signatures because they get repositioned when
imported. However, sometimes errors are reported without repositioning.

Adding the necessary repositioning still makes sense as a longer-term strategy,
since it helps localize errors. However, for now we take the conservative
approach of *not* assuming that repositioning always kicks in, thereby including
all locations reachable from exports as part of the sig hash.

Reviewed By: gabelevi

Differential Revision: D4403313

fbshipit-source-id: f2f9bcbb6c4e9b81b9f77228a3e0289630fad0ee