improve object compatibility errors when a prop only exists in one obj
Summary:
when two objects are unified, flow errors if one object contains a
property that the other doesn't. we have a `use_op` that is keeping track of
which objects are being compared while we compare each prop. this diff threads
it through to the error, so we can error on the objects instead of the props.

Reviewed By: gabelevi

Differential Revision: D4404859

fbshipit-source-id: 923f7595eb34b48ada51e1bfa3c5907845e73d66