improve object compatibility errors when prop types mismatch
Summary:
when object types are unified, we unify the types of each common prop.
this diff creates a `PropertyCompatibility` `use_op` for this and threads it
around, which causes the errors to point at the objects being unified instead of
the types of the props that are incompatible.

note that we already do this when flowing an object literal into an object
type/literal. this is specifically when two object types collide.

Reviewed By: gabelevi

Differential Revision: D4405587

fbshipit-source-id: 25da51af3e0b46eaa21d585c818825034db18ffb